honestly, I probably broke everything but maybe I'll fix it  eventually?
