+++ src/static/script.js
