+++ src/static/src/input.css
