+++ src/templates/del_msg.html
