+++ src/templates/panel.html
